pyproject.toml
[tool.poetry]
name = "object-api"
version = "0.0.1"
description = ""
authors = ["Jacob Valdez <jacobfv123@gmail.com>"]
readme = "README.md"
packages = [{ include = "object_api" }]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.24.1"
black = "^23.7.0"
fastapi = "^0.101.0"
sqlmodel = "^0.0.8"
pydantic = ">=1.8.2,<2.0.0"
inspect-mate-pp = "^0.0.4"
stringcase = "^1.2.0"
scheduler = "^0.8.4"
glom = "^23.3.0"
python-exports = "^1.1.0"
uvicorn = "^0.23.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
